% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcctz-package.R
\docType{package}
\name{rcctz-package}
\alias{rcctz}
\alias{rcctz-package}
\title{rcctz: Vendor 'CCTZ' Headers}
\description{
Provides access to the 'CCTZ' headers. 'CCTZ'
    contains two libraries that cooperate with the C++ 'chrono' to give C++
    programmers all the necessary tools for computing with dates, times,
    and time zones in a simple and correct manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/rcctz}
  \item Report bugs at \url{https://github.com/DavisVaughan/rcctz/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
